Numeric fix for 1.8 (#1711)
* cherry pick no compression on numeric for 1.8

* change default when used by as library

+++ src/numeric_index.h
@@ -53,7 +53,7 @@ typedef struct rtNode {
 } NumericRangeNode;
 
 typedef struct {
-  size_t sz;
+  int sz;
   int changed;
 } NRN_AddRv;
